server/requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.11.1
click==8.1.3
Deprecated==1.2.13
Flask==2.3.2
gunicorn==22.0.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
redis==4.5.4
soupsieve==2.3.2.post1
Werkzeug==3.0.3
wrapt==1.14.1
zipp==3.10.0
mysql-connector-python==9.0.0
